Missing thumbnails are now created at System Helper launch

+++ system_helper/helper.py
@@ -603,22 +603,14 @@ def delete_file(file: str, absolute: bool = False):
     with config.content_file_lock:
         os.remove(file_path)
 
-    mimetype, _ = mimetypes.guess_type(file)
-    mimetype = mimetype.split("/")[0]
-    if mimetype == "image":
-        thumb_path = get_path(["thumbnails", with_extension(file, "jpg")], user_file=True)
-    elif mimetype == "video":
-        thumb_path = get_path(["thumbnails", with_extension(file, "mp4")], user_file=True)
-    else:
-        thumb_path = None
-    if os.path.exists(thumb_path):
+    thumb_path, _ = get_thumbnail(file)
+    if thumb_path is not None and os.path.exists(thumb_path):
         with config.content_file_lock:
             os.remove(thumb_path)
 
 
 def get_all_directory_contents(directory: str = "content") -> list:
     """Recursively search for files in the content directory and its subdirectories"""
-
     content_path = get_path([directory], user_file=True)
     result = [os.path.relpath(os.path.join(dp, f), content_path) for dp, dn, fn in os.walk(content_path) for f in fn]
 
@@ -888,6 +880,24 @@ def with_extension(filename: str, ext: str) -> str:
     return ".".join(split[0:-1]) + "." + ext
 
 
+def get_thumbnail(filename: str) -> (Union[str, None], str):
+    """Check the thumbnails directory for a file corresponding to the given filename and return its path and mimetype"""
+
+    mimetype, _ = mimetypes.guess_type(filename)
+    mimetype = mimetype.split("/")[0]
+    if mimetype == "image":
+        thumb_path = get_path(["thumbnails", with_extension(filename, "jpg")], user_file=True)
+    elif mimetype == "video":
+        thumb_path = get_path(["thumbnails", with_extension(filename, "mp4")], user_file=True)
+    else:
+        thumb_path = None
+
+    if thumb_path is not None and not os.path.exists(thumb_path):
+        thumb_path = None
+
+    return thumb_path, mimetype
+
+
 def create_thumbnail(filename: str, mimetype: str):
     """Create a thumbnail from the given media file and add it to the thumbnails directory.
 
@@ -923,6 +933,16 @@ def create_thumbnail(filename: str, mimetype: str):
         except OSError as e:
             print("create_thumbnail: error:", e)
 
+
+def create_missing_thumbnails():
+    """Check the content directory for files without thumbnails and create them"""
+
+    content = get_all_directory_contents("content")
+
+    for file in content:
+        file_path, mimetype = get_thumbnail(file)
+        if file_path is None:
+            create_thumbnail(file, mimetype)
 
 def get_path(path_list: list[str], user_file: bool = False) -> str:
     """Return a path that takes into account whether the app has been packaged by Pyinstaller"""
@@ -1047,9 +1067,11 @@ if __name__ == "__main__":
 
     signal.signal(signal.SIGINT, quit_handler)
 
-    # schedule = None # Will be filled in during read_default_configuration() if needed
-    # NEXT_EVENT = None
     read_default_configuration()
+
+    # Check for missing content thumbnails and create them
+    create_missing_thumbnails()
+
     # If it exists, load the dictionary that maps one value into another
     load_dictionary()
